brain/pyproject.toml
[tool.poetry]
name = "pluto-brain"
version = "0.1.0"
description = "LangChain-powered brain service for Pluto"
authors = ["Pluto <dev@pluto.local>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.109.0"
uvicorn = { extras = ["standard"], version = "0.26.0" }
python-dotenv = "1.0.1"
langchain = "0.1.5"
langchain-openai = "0.0.5"
httpx = "0.26.0"
SQLAlchemy = "2.0.25"
asyncpg = "0.29.0"
pydantic = "2.6.1"
pydantic-settings = "2.2.1"
tavily-python = "0.3.4"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.4"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"
